Invoke isort independently on running lint target
Pylama is configured to invoke isort to lint import statements already.
As a result, it prints the files where we need to fix the import
statements. But it does not tell us in detail exactly what changes we
need to make to fix the import statements.

Therefore, in this change, we invoke isort independently to show us the
changes (in diff format) that we need to make to fix the import
statements.

This independently invoked isort exits with exit code 0 regardless of
whether it finds problems with import statements or not. So it does not
prevent the subsequent Pylama command from running even if it finds
problems with the import statements.