Add unit tests for MongoDBStore
The unit tests use `mock.patch` to patch `MongoClient` with `MagicMock`
objects in order to avoid dependency on an actual MongoDB server to run
the tests.

Note that in the tests we do not call a `Mock` object's `assert_called`
method although this method would have sufficed. That's because
`assert_called` method was introduced in Python 3.6. We support Python
versions 3.4 and 3.5 also at this time and therefore run Travis CI
builds with these versions. Calling `assert_called()` leads to test error
with Python 3.5 due to `AttributeError` and does not do what we want on
Python 3.4 (returns a `MagicMock` object that does nothing).

In order to support Python versions 3.4 and 3.5, we use
`assert_called_with` and `assert_called_once_with` methods because these
methods are available prior to Python 3.6.