Add AzLogProfileMissingCategoryEvent plugin
A log profile controls how the activity log is exported. Configure the
log profile to collect logs for the categories `write`, `delete` and
`action` ensures that all the control/management plane activities
performed on the subscription are exported.

The `AzLogProfileMissingCategoryEvent` plugin evaluates records of
type 'log_profile' for Azure cloud and generates event of type
`log_profile_missing_category_event` if a log profile is not
enabled for all three activity types i.e. `Write`, `Delete` and
`Action`.